+++ src/PSCompression/Commands/ConvertFromBrotliStringCommand.cs
@@ -1,7 +1,6 @@
 using System.IO;
 using System.IO.Compression;
 using System.Management.Automation;
-using BrotliSharpLib;
 using PSCompression.Abstractions;
 
 namespace PSCompression.Commands;
@@ -12,5 +11,5 @@ namespace PSCompression.Commands;
 public sealed class ConvertFromBrotliStringCommand : FromCompressedStringCommandBase
 {
     protected override Stream CreateDecompressionStream(Stream inputStream) =>
-        new BrotliStream(inputStream, CompressionMode.Decompress);
+        new BrotliSharpLib.BrotliStream(inputStream, CompressionMode.Decompress);
 }